move clients building logic into ruby_block resource
This is needed for when a wrapper cookbook modifies the client's backup
strategies during runtime, instead of compile time.

An example of this would be something like this:

    ruby_block 'manage_backup_strategies' do
      action :create
      block do
        Backupable::DefaultStrategy.new(node)
        Backupable::BaseStrategy.new(node) if node.recipe?('base::default')
      end
    end

the `DefaultStrategy` `BaseStrategy` classes would set the node attributes
`node.bpc.client.includes` and `node.bpc.client.excludes`. This needs to be in
a `ruby_block`, because the expanded run-list is unavailable during compile
time, and thus `node.recipe?` might not return true if the  `base::default`
recipe hasn't been compiled yet by Chef.